requirements.txt
openai>=1.30.0
colorama>=0.4.6
python-dotenv>=1.0.0
gradio>=4.0.0
sentry-sdk>=2.0.0
